+++ include/IsoVerifierWidget.h
@@ -4,6 +4,7 @@
 
 #include <QWidget>
 
+class QComboBox;
 class QFrame;
 class QLineEdit;
 class QPushButton;
@@ -13,6 +14,7 @@ class QLabel;
 class QProgressBar;
 class QTextEdit;
 class QSplitter;
+class QStackedWidget;
 
 namespace FlashSentry {
 
@@ -33,10 +35,12 @@ public:
                           const QString& deviceLabel = {});
 
     void refreshCatalogStatus();
+    void setActiveProfile(const QString& profileId);
 
 signals:
     void logMessageRequested(const QString& message);
     void verificationReportReady(const QString& deviceNode, const QList<IsoVerifyResult>& results);
+    void settingsProfileSelected(const QString& profileId);
 
 public slots:
     void onUsbMountForIsoVerify(const QString& mountPoint, const QString& deviceNode,
@@ -51,20 +55,28 @@ private slots:
     void onUpdateCatalog();
     void onPasteTrustedHash();
     void onScanPathEdited();
+    void onProfileChanged(int index);
+    void onTableCellClicked(int row, int column);
     void onVerificationFinished(const QString& location, const QString& deviceNode,
                                 const QList<IsoVerifyResult>& results);
 
 private:
+    QWidget* buildEmptyStatePage();
+    QWidget* buildMainPage();
     void setResults(const QList<IsoVerifyResult>& results);
     void appendReport(const QString& text);
     void updateCatalogIntegrityBanner();
     void updateMultibootBadge();
     void updateSummaryStrip(int passed, int total, int needsSidecar);
+    void updatePageVisibility();
+    void scrollReportToRow(int row);
     void applyChromeStyles();
     void styleResultRow(int row, bool passed);
 
     IsoVerifierWorker* m_worker = nullptr;
+    QStackedWidget* m_pageStack = nullptr;
     QLabel* m_catalogBanner = nullptr;
+    QComboBox* m_profileCombo = nullptr;
     QLabel* m_introLabel = nullptr;
     QLabel* m_multibootBadge = nullptr;
     QFrame* m_summaryStrip = nullptr;